+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 2020-04-28
 * Added Contact Preferences extension to the Patient resource
+* Removed GET Related Person by Object ID
 
 ## 2020-04-27
 * Add new regression and smoke test packs